added threads page
