+++ frontend/package.json
@@ -32,6 +32,7 @@
     "next": "13.2.4",
     "next-auth": "^4.20.1",
     "postcss": "^8.4.21",
+    "query-string": "^8.1.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "tailwindcss": "^3.3.1",
