+++ frontend/src/pages/api/useCreateClass.ts
